authentication-service/requirements.txt
Flask
PyJWT
Flask-HTTPAuth
psycopg2-binary
